Update privacy policy
Updated the privacy policy document to reflect changes in data handling methods, new applications added, and services removed. Updated the date of last modification and added a section to clarify instances when no data is stored on our servers and when data is stored on third-party servers, to maintain transparency with our users.